renaming and moving files temporarily to try and simplify the merge conflict
