+++ src/components/Editor/Runners/RunnerFactory.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import PythonRunner from "./PythonRunner/PythonRunner";
+import PythonRunner from "./PythonRunner/NewPythonRunner";
 import HtmlRunner from "./HtmlRunner/HtmlRunner";
 
 const RunnerFactory = ({ projectType }) => {
